refactor(final_step): rename module json_format.py to json.py
